match: full-toolbox sweep harvest on flag-qualified near-misses
Wins from the pattern-sweep-harvest workflow (full permuter on db-flag-qualified candidates), all verified via objdiff full_build:

MoveMgr::ComputePotentialMoves 91.6->93.5 (member_ref_bind), RndText::UpdateScrollOffsets 98.4->99.2 (switch_case_reorder), Trie::dec_dup_count 94.8->95.2 (variable_inline), BuildSphereStratified 99.8->99.9 (varext+declreorder), StreamReceiver::Poll 91.6->91.8 ((unsigned)mState compare — matches existing line-100 idiom).

Rejected matcher-gaming variants: HamLabel::Count argument_swap (made Key<float> ctor args inconsistent), HollaBackMinigame::OnBeat null_guard_insert (adds control flow), Utl::XfmSort comparison_flip on a sort comparator (also non-reproducible).

+++ src/system/synth/StreamReceiver.cpp
@@ -101,7 +101,7 @@ void StreamReceiver::Poll() {
         if (mState == kReady) {
             goto ready;
         }
-        if (mState > kStopped) {
+        if ((unsigned int)mState > kStopped) {
             MILO_FAIL("bad state logic.\n");
             goto ready;
         }
